Add comments to Behavior Record
Comments are allowed in Behavior Record.
The RodentVR adds a comment at the top of the file
in order to specify what each data field is.
The installer packages were updated to reflect
these changes. In addition, the installer for
RodentVR was modified to create smaller files which
can be added to git.


Former-commit-id: e013f0f28b45a0acbd38c9766379da246e6dee4a
Former-commit-id: 8721b2f31b0e6ae03667ec044c4e1de08694dfd4