Change welcome routine, add new config
Now, when the user starts the program, they receive
a welcome message that asks them whether to create
a new configuration file or to open an existing one.
This way, the user is fully aware of what is happening.

Also, add a button to the main menu that allows the
user to create a new config file. Warn the user before
that it will erase an unsaved changes.


Former-commit-id: b1b17c45f7402bd8b4e7ca95b52803ac4072e785