Improve UI Appearance
I changed all the widget in (almost) the entire program to inherit from
base widgets that I stylized. That way fonts, padding, sizes, etc would
be standardize across the program. That allowed me to easily experiment
with themes and pick a theme that looks pretty cool. If I changed the
base style, everywhere it was used in the app would change, which is
handy.


Former-commit-id: a60a79d5d97f36b8e24200bc3dc7e843fd4885b6
Former-commit-id: d95259be1d3275cf1cf6f5726e1ee2aba7fd1627