Remove Unintentionally Committed Files
Back when I first created the project, I didn't have a .gitignore for
UE4 (or atleast did it wrong...) So I have 9575 generated files
(actual number) that I didn't want in the repo. I just found out that I
can remove them so as to respect the gitignore now. This will clean up
the repo quite a bit! Sorry to anyone who inherits this project, I'm
still learning the ins and outs of UE4, and messed up with UE4 and git!